Move move menu component from mobile to template folder
