+++ src/lib/components/organisms/mainFooter.svelte
@@ -2,7 +2,7 @@
 	import { invalidateAll } from '$app/navigation';
 	import { page } from '$app/stores';
 	import MobileOnlyFooter from '$lib/components/zMobile/mobileOnlyFooter.svelte';
-	import MoveMenu from '$lib/components/zMobile/moveMenu.svelte';
+	import MoveMenu from '$lib/components/templates/moveMenu.svelte';
 	import { selected } from '$lib/stores';
 	import { isSelectModeOnMobile } from '$lib/stores';
 	import { trpc } from '$lib/trpc/client';
